1. Хранилище продуктов [#852]

+++ src/main/java/ru/job4j/ood/lsp/products/food/Meat.java
@@ -2,8 +2,8 @@ package ru.job4j.ood.lsp.products.food;
 
 import java.util.Calendar;
 
-public class Meat extends Food{
-    public Meat(String name, Calendar expiryDate, Calendar createDate, double price, int discount) {
+public class Meat extends Food {
+    public Meat(String name, Calendar expiryDate, Calendar createDate, int price, int discount) {
         super(name, expiryDate, createDate, price, discount);
     }
 }
